[fuchsia-async] Reimplement TaskGroup in terms of Scope
As a next step we can remove TaskGroup entirely, but I wanted to swap
out the backend as a canary for implementation issues and so we can
announce zxdb support for scopes.

This required adding ScopeRef::spawn_local and while we're at it,
ScopeRef::compute_local APIs.

This removes the TaskGroup::add method since it wouldn't be very
efficient to implement. We might want a reparent method eventually but
shouldn't encourage it as a way to spawn tasks.

Bug: 339724492
Change-Id: I5e70c769d049dd8744e2419d1d12611585f936d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148332
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>