[sdk] Revert zircon-internal to a standard source_set()
https://fxrev.dev/979816 converted "zircon-internal" to an
`sdk_source_set()` because it was a `public_deps` of "fake-object", but
it is not actually used by "fake-object".

This reverts changes to zircon-internal in commit f6a44a00adb4da2219f225456d91adbb8381c9f6.

Remove that dependency and correct dependencies in "devicetree" so that
there is not an (SDK) public dependency from that SDK static library.

Bug: 372986936
Change-Id: I5e0d927127b41ac62c5c91461a9a4213640b1a67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149581
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>