[test] Require filename when using ConstantMetricsProcessor to report freeform metrics
The filename under which freeform metrics are saved is part of the
publicly exposed information when the metrics are stored in our backend.
As such, require that callers using the ConstantMetricsProcessor to push
freeform metrics into the system pass an explicit identifier to be used
when publishing metrics.

Bug: b/376922110
Test: trace_metrics_test cpu_breakdown_test
Change-Id: Ic2a45c416fe86ad31399e15382f161a2ebc786ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149474
Commit-Queue: Chris Masone <cmasone@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>