[Honeydew] re-enable input and session tests on internal vim3 builder
Session tests: For tests that require the session to be in a healthy
state before the next action, we use a timeout so that all
session-associated components can finish starting up or shutting down
fully.

Input tests: The screenshot data doesn't properly reflect the app
pixels, so we remove the assertion as was done for the screenshot test.
See b/320543407 for more details. It is still worth enabling this test
without the screenshot functionality so we can exercise touch device
creation and tap injection in the meantime.

Fixed: 370481460
Change-Id: I6f3ad7fc618d3932d502f9e86a185e27087fadc2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137593
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>