[sestarnix] Set xattr on FsNode, if supported, on creation
When a new node (file, directory, symlink, etc.) is created in a
filesystem, an in-memory label is determined for it based on the
labeling scheme configured by the loaded policy.  If the filesystem
uses "fs_use_xattr" labeling then the label is also serialized and
returned to the Starnix VFS, to write to the new file node.

While in principle no xattr should be returned unless the
filesystem is configured with "fs_use_xattr" labeling, use of the
"fscreate" task attribute when creating files in non-xattr-capable
filesystems requires that ENOTSUP is silently ignored, for now.

Bug: 355180447, 376429620
Change-Id: Ie8ec41e0e6a49f57ffb46845af1d68332b033fbc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1150336
Reviewed-by: Nelly Vouzoukidou <nellyv@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>