[symbolizer] Fix Negative Loads with Multiple Modules
Zircon on x64 has negative loads, which is accommodated by treating its
base address as 0. This caused a conflict when searching through the
modules looking for an address by upper/lower bound because it would
find a different module and find that the zircon address is not
contained in the module.

This works for symbolizing kernel backtraces right now because in every
instance, the zircon module info is the only info that's provided, so
the upper/lower bound just happens to find it.

This patch adds a fallback to check for a module with 0 base address if
the wrong module is found.

Bug: 377355351

Change-Id: Ibc6c99609ef5ab872fead8228f21702cf101d4b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1150712
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>