[network/bench] Shorten system metrics trace duration
The iPerf benchmarks rely on the system metrics daemon to collect CPU
usage stats and output them as trace events. In order to ensure that the
daemon is collecting data during the benchmark, the benchmark records a
short trace session before beginning and checks if it includes any CPU
usage trace events. The duration of this trace session is currently 10
seconds, which is longer than typically needed to capture some CPU usage
records. Shorten it to 1 second so that even if we have to retry
occasionally to get a CPU usage event, we don't always take at least 10
seconds to start the benchmark.

Bug: 370786547
Test: netstack_iperf_test
Change-Id: I279186c3ae338059ee430a28ad456ed89b85e1fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143532
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>