[netlink] Introduce IndexedRules data structure
Introduces an IndexedRules data structure that stores a set of rules
that all have the same priority. Internally, this is a
BTreeMap<u16, RuleMessage>, where the u16 is the rule's index among the
set of rules with the same priority.

This replaces what was previously a Vec<RuleMessage>, which will be
unsuitable once we start installing rules via the relevant FIDL
protocols, which expect the "index" of a given rule to remain the same
as long as the rule is installed.

Now, when a rule is removed, all the rules after it don't have their
indices changed.

Test: netlink-lib-test
Bug: 358649849
Change-Id: Ie43be2d46ee48bf5e1968a70cbc338a8c2d49f15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148072
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>