[netstack3] Enter ZWP on zero window update
As written, if the window shrinks to 0, but no new data is acked, then
the retransmit timer won't be cancelled. If there's a timer set, then
Netstack3 doesn't enter ZWP. This leads to a situation where it
continually tries to retransmit, but since the window is zero, it can't
send any data.

Fixed: 377294231
Change-Id: I9b7f15b538f5a80ed03cc4656df5b16b89f3d75b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1150293
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>