[vim3] Switch back to using 16codegen units on debug builds
The use of a single codegen unit was a time/space tradeoff that only
bought about a month before the zbi was again oversize, and the zbi
compression level was increased to make the zbi smaller.  This higher
zbi compression level allows the use of 16 codegen units again, and
allows us to trade 15 seconds of time compressing the zbi vs minutes
of compilation spent using a single codegen unit.

Change-Id: I3265025274f868274909d30f57020859a24bb537
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151113
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>