[network/bench] Run iPerf for 5s instead of 10
Currently the iPerf benchmarks (both the host-driven ethernet and
hermetic loopback variants) run iPerf for the default 10-second
duration. The tests are currently taking up a lot of time to run in CI
and we can get results that are nearly as useful by running it for just
half the time, so reduce the runtime to 5 seconds.

From running perfcompare, the effects are this are likely to be:
- slightly lower CPU usage for multi-flow ethernet tests
- very slightly lower throughput on the ethernet receive tests
- fewer lost packets for UDP loopback netstack3 tests

For more analysis of why these changes are unconcerning, see
https://fxbug.dev/370786547#comment3.

Bug: 370786547
Test: netstack_iperf_test
Change-Id: Id1776f76225147d16df1fb800d54aaf8f08b626a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143533
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>