[selectors] Fix validation of string patterns
We were requiring some chars to be escaped when it was unnecessary for
htem to be escaped. The only thing that should require escaping in a
string pattern is the `*` when used as `**`.

Escaping of `:` and `\` is necessary at the selector language level, but
it's not necessary when we are just dealing with string patterns. We can
still escape `*` in string patterns, as that has a different meaning
than a raw `*`.

This is similar to how we unescape exact matches. We should have been
unescaping string patterns as well, taking into account the edge case of
`**`.

Change-Id: Ib292b5fffac7e75e78fb16a4dab8e0954936ac67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151239
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>