[ftl] Start tracking wear with uint32.
Stops tracking the values as "lag" from the high wear either, uses them
as raw values.

Removes a bunch of statistics that are based around the wear lag limit,
and defers some calculations to the stat retrieval or debug printing
rather than shouldering the complexity of keeping them up to date.

Leaves the wear lag histograms in place to appear the same as
previously.

Bug: b/376287044

Change-Id: I801fd75ac570bfac8f9dfb068b82d3bb22e9c0bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147974
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>