[network/bench] Start iPerf server only once
Start the on-device iPerf server once at the beginning of the benchmark
and reuse it across test cases. This does have the downside of server
output not being saved separately for each test case, and a server
failure in one test case causing issues in a subsequent one, but it
saves a lot of overhead on server setup and teardown that currently
applies for every test case rather than once for the entire benchmark.

Also simplify the code now that we only run single-flow benchmarks over
ethernet; we no longer have to handle running multiple clients and
servers in parallel outside of the hermetic benchmark.

Bug: 370786547
Test: netstack_iperf_test
Change-Id: I3be719fe7c7f72de2c1ecfefe6366a152f3d0286
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1149536
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>