[inspect] track "at capacity" duration of history
This patch adds a specific "duration of history"
stat for when the buffer is running at capacity.

The earlier "history_duration_ns" is useful for
humans but because it also measures duration in
the early phase of the buffer, it interferes with
Cobalt MIN_PROCEDURE sampling. This variant is
vended only when the buffer is at capacity.

Also, adjust inspect type to "int" to better match
the zx duration type.

Bug: 365600132
Change-Id: I4a6c3d3073bd9ff6909be8b2ce904becef71ecb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151036
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Jaeheon Yi <jaeheon@google.com>