[paver] properly report last-boot-attempt
The firmware sets a slot to "unbootable" just before the final boot
attempt, but this caused some unexpected behaviors in the OS when the
current boot slot is reporting itself to be unbootable.

This CL adds detection for this case, and adjusts the paver FIDL data to
properly report the slot as pending but with zero attempts left.

Bug: b/368597963
Change-Id: Ie81d353730a3e6bc0a957f05c2dff40c8f172adf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148338
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
API-Review: David Pursell <dpursell@google.com>
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>
Commit-Queue: David Pursell <dpursell@google.com>