[zxio] Reduce Remote template verbosity
Reduce verbosity in using the Remote type by replacing the use of a
constant value for the zxio object type with a constexpr function that
derives the same information from the protocol itself. The constant was
only ever used for a single zxio operation (attr_get), so by removing it
as a template argument we improve readability without sacrificing
maintainability.

Test: fx test //sdk/lib/zxio
Change-Id: I8e482b33f99b87038e407ea475cfae2c7313e8aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151572
Reviewed-by: Rui Qi Sim <rqsim@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>