[docsgen] Fixing clidoc dependencies
This uses the sdk_molecule manifest as input to determine
the list of tools to run to generate the reference docs.

This is done to make sure all the tools are dependencies of
the invoke_clidoc target to make sure the build is consistent.

Bug: 376723774
Change-Id: I8b582264fa33b70b37f068eda65bf2b15e8aadd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1148896
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>