[display] Move utils to standalone library.
This CL does not introduce any new code. All utilities are moved from
util/ directories in existing drivers. The drivers themselves are
modified to use the headers and classes/functions in the library.

Updating driver usage includes adding the display:: namespace qualifier
to the classes and functions that are now in the library. This accounts
for the vast majority of the driver changes.

Bug: 359960626
Test: fx test //src/graphics/display/lib/driver-utils
Test: fx test //src/graphics/display/drivers
Change-Id: I8082ea8f11eae5bfc1a225925358638c803976d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151373
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>