[bazel][clang] compile includes runtime libdirs
In the referenced bugs, it was found that clang's driver
is sensitive to the presence of runtime libdirs when computing
the set of include dirs used for compiling (cc1).
The absence of the said runtime libdirs would cause the driver to
omit the correct multilib variant of headers like __config_site,
which would then cause a mismatch of variants (e.g. ASan vs. non-ASan)
downstream when linking, and sometimes not manifesting problems
until runtime.
This workaround helps the driver select the correct multilib headers.

Bug: 354016617
Bug: 360235447
Change-Id: I07b28e53b957fa347c86cb0a5ce6919432d8c4f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1150592
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>