Reland "[docsgen] Fixing clidoc dependencies"
This is a reland of commit 87ebfa6dcee9f19e8a32e6d0ace1a4c40cd4f128
The revert was due to a doc change that was a problem which slipped in,
so unrelated to this CL.

This uses the sdk_molecule manifest as input to determine
the list of tools to run to generate the reference docs.

This is done to make sure all the tools are dependencies of
the invoke_clidoc target to make sure the build is consistent.

Bug: 376723774
Change-Id: Ib2948d082727e104d7c344dff1d6c1486b42e1fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152353
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>