[tools/test_pilot] add parameter intake module
Adds a module for processing test parameters.

This CL adds a module that handles the intake of test parameters (and
some other command line options) via the command line, environment
variables, and JSON files. This is part of a exploratory implementation
of the new testing ABI. Future design changes might require this code
to change considerably.

The approach taken here is based on two tenets:

1. The way test parameters are delivered to test_pilot (via command
   line, environment variables or JSON files) should be independent
   of the actual name/types/purposes of the test parameters themselves.
   This means that any given parameter may be submitted in any of these
   three ways (with some exceptions).
2. test_pilot should only understand those test parameters that are its
   direct concern. Any consistency checks it performs outside of those
   parameters should be driven by constraints expressed in the test
   parameters themselves. This would take the form of e.g. parameter
   declarations, `required` lists, `prohibited` lists, etc.

The new module includes a debugging facility that prints a detailed
report of the test parameter processing performed by the module.

Bug: 294567877

Test: new unit tests
Change-Id: Ic7ce1c0342fd447c43e5ee8cd5303746c4455ab8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140099
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>