[tee][api] Stand up persistent storage interface scaffolding
This change stands up scaffolding for the implementation of the subset
of the trusted storage interface dealing in "persistent". In this first
pass, the lifetimes of these objects are scoped to the individual TA
instance, rather than to all instances.

Bug: 360942417
Change-Id: I33a9e9ee99a17cbbb30318c9814575083a3bc241
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1151512
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>