[kernel][vm] Expanded attribution counts
These allow implementing the APIs in RFC-0254.

None of the new count fields are exposed to userspace yet. As specified
in the RFC, the new integer fields are set to 0 and the new fixed-point
fractional fields are set to UINT64_MAX.

We do implement the logic to internally sum these new counts across VMOs
and address spaces, even if we don't exposed the summed values to user
space yet.

No functional change.

Bug: 42058561
Bug: 338300808
Change-Id: Icb67e4284134e0e1442094bcc2e8761ec8d0249a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147212
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>