[fuchsia-async] Remove WeakScopeRef, Join::detach()
As the documentation explains, the type is unnecessary because it can't
be used to create cycles that didn't already exist. A ScopeRef can be
held by a user's task code, but each Task already holds a ScopeRef, so
it would just be redundant.

WeakScopeRef does prevent cycles in the internal ownership graph.

Testing this also exposed a flaw in the Join API. The intention was to
allow cancelling or detaching after polling the Join future, but Join is
not Unpin, so those methods must accept `Pin<&mut Self>` to be useful.
Admittedly this is a clunky API to use. I fixed cancel() but removed
detach() since implementing it would require putting the inner scope
in an Option and for the rare cases that would ever need it, mem::drop()
is a way out.

Bug: 339724492
Change-Id: Ibc9082bd15dd0f693c64613127f33a5a28189afc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152733
Fuchsia-Auto-Submit: Tyler Mandry <tmandry@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>