[fuchsia-async] Add ScopeRef::cancel()
This allows cancelling a scope when it is not owned. This might be
useful for cancelling a scope from within a task spawned on that scope,
for example.

Now Scope only differs from ScopeRef in that it

1. Cancels the scope automatically on drop.
2. Allows joining the scope (waiting for no tasks + closing).

Joining the scope could also be allowed from a ScopeRef, but as I
explain in the code comment, this seems less of a good idea.

Conceptually, the Scope/ScopeRef relationship is aligned with the
Task/JoinHandle relationship in that the first "owned" variant cancels
the object on drop. We should probably align the naming of these.

Bug: 339724492
Change-Id: Ia54ae364eaad2558fd0240be0a09f7fd7c477920
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152812
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>
Fuchsia-Auto-Submit: Tyler Mandry <tmandry@google.com>