Reland "[orchestrate][repo] Orchestrate adopting standalone package server"
This is a reland of commit 8b2b5d129ba94942c80c1b44f1b3a521cb66cf01

The revert was due to a broken dependency in //tools/docsgen which was addressed in fxr/1148896.

This changes orchestrate to use the standalone repository package server.

Bug: 374085877
Change-Id: Id63135503a802f2f2b8ac2a16068ca0e2bf5321f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152331
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>