[sestarnix] Add permissions checks to fs_node_setsecurity()
When a file node's "security.selinux" xattr is modified, the system
validates that the task has "relabelfrom" and "relabelto" rights to
the new and old Security Contexts, respectively.

The new Security Context should also be checked for the "associate"
permission to the file-system; this check is made, but the result
ignored, until correct file-system labels are implemented.

Bug: 351195217, 355180447
Change-Id: I9f80f3064da711117d639adfb63de9b6880013ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112132
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nelly Vouzoukidou <nellyv@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>