[session] Remove builtin runner + boot resolver session-env
We route some capabilities to the session that it does not use, and are
only used to launch platform components. To protect against an unwanted
dependency in the future, remove them from the session's environment.

These capabilities are:

- boot-resolver: For launching fuchsia-boot components
- builtin_runner: For launching builtin components (components in
  component_manager). An example is the elf runner component.

In the case of #2, we are adding more builtin components, which were
designed only for platform use.

Fixed: 375000535
Change-Id: I484f7dda8673c5daaff123f999934299ef567bff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1146981
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Gary Bressler <geb@google.com>