[build] Enable frame pointers by default in release.
The absence of frame pointers in release builds is the only
obstacle to being able to run performance profiling tools like
the CPU profiler or heapdump on "regular" release builds. It
would streamline workflows and reduce the number of bot
configurations we need to be able to use the default release
artifacts for profiling.

This does increase the size of the system image, but it is a
dependency of the project(s) to remove unwind tables which should
eventually net us a very significant binary size reduction.

Previous toolchain investigations have shown that the performance
overhead of frame pointers is negligible.

Bumps several core size limits to ensure that there is at least
as much headroom as a single instance of each bucket's creep
budget.

Bug: 42164648
Bug: 42053657
Bug: 369225659
Bug: 359658592

Change-Id: I222f7603d2611aff3900072385345dbb5f7942e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147933
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Size-Review: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>