[fuchsia-async] Rename ScopeRef to ScopeHandle
And document the scope lifecycle better.

I haven't liked the "Ref" naming that much, since it evokes the idea of
&T, so I prefer "Handle".

JoinHandle is kept the same because of its similarity with the ones in
tokio and in std::thread. It is much more limited in its capabilities
compared to a ScopeHandle, so renaming it to TaskHandle by analogy to
ScopeHandle (as this CL originally did) suggests more similarity than
there actually is.

NO_IFTTT=ScopeRef does not appear by name in zxdb source

Bug: 339724492
Change-Id: I43bc36532b584e520d509c9f71722b0a4a37d21a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152852
Commit-Queue: Tyler Mandry <tmandry@google.com>
Reviewed-by: Chris Suter <csuter@google.com>