[hrtimer][vim3] Add setup_event to StartAndWait
The `setup_event` event signals to a client that the driver has
setup the timer in the hardware. Hence for cases when the
system may need to wakeup from this timer, it is safe to allow the
system to suspend now that it is known that the timer is setup.

Bug: 373928684
Test: fx test -o aml-hrtimer-test
Test: Exercise a timer on VIM3:
$ hrtimer-ctl -i 0 -w 1
Executing on device /dev/class/hrtimer/27021865
Starting timer...
Timer triggered
Keep alive released
Timer started

Change-Id: I67e06bb3c9300f3d9a0a2672252eb5e62ebfe28d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1147532
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
API-Review: Michael Brunson <mbrunson@google.com>