[fxl][log] Expose CLI parsing without the syslog dep
The json_validator is used as part of the FIDL toolchain. It uses the
FXL command line parsing utilities. Syslog would like to be able to use
log levels defined in FIDL as the source of truth for log levels.
There's optional support in the FXL command line parsing utilities to
parse log levels. This causes a cycle.

This introduces a new cli_no_logging target to FXL that's just the
non-log-related parts of command-line parsing. This was already written
in separate files and is all that json_validator actually uses.

Change-Id: I2c7c2aca46ac0454b5f607eece75becda6b24e59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152516
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>