[netstack3] Add switch for sending test data
Many tests use the bind_listen_connect_accept_inner function to set up a
connected pair of sockets, but not all of them rely on the test data
that gets sent as part of that process. However, for an upcoming test
case, this test data breaks the test because it affects the advertised
windows.

Adding this flag required auditing the test cases to see whether they
actually relied on the data being sent, which for the most part they
don't.

Bug: 372763809
Change-Id: Id328912866e128e6f9cddbbac4410437a7147ae4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153712
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>