[logs] Fix error with interest setting when reconnecting to a device
Previously, when a device rebooted, if an interest level
was specified, the reconnect would fail due to the proxy being
closed, because the proxy was passed outside the general connection
object. Now, the proxy is part of the connection object so
this should no longer happen.

Fixed: 368353859
Change-Id: I5b1c56fa53a5d87b7ce39d690917fe53347e7c04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153818
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>