[elfldltl] Replace NewArrayFromFile with ContainerArrayFromFile
This eliminates a dependency on fbl::AllocChecker in the
widely-used <lib/elfldltl/memory.h> public header.

The only users of NewArrayFromFile are in test code and thus
don't need to be fully optimized.  The replacement for that class
is a helper function that adapts Container API object to the
Allocator API for ReadArrayFromFile in the Memory API.  This is
suboptimal because the container API has no way to allocate space
without default-constructing each object, whereas the old code
using `std::unique_ptr<T[]>` could allocate uninitialized space.

An optimized replacement would do something more like the old
NewArrayFromFile did, but could perhaps be separated into a
different public header to isolate the fbl dependency and/or
template-parameterized somehow to handle both plain new and
AllocChecker new.  But that can be revisited when a need like
that arises.  So far the only performance-critical uses of
ReadArrayFromFile (startup and libdl loaders for phdrs) just use
FixedArrayFromFile.

Change-Id: Ibf3799570441691e0676ff06e60ab9be442361ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152874
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>