[network/bench] Read iPerf results from file instead of logs
Have iPerf write its results to a file in /custom_artifacts rather than
attempting to retrieve them from syslogs, as it's possible for Archivist
to either process the log lines too late for the benchmark to observe
them or to drop them altogether.

This requires adding a custom proxy for custom_artifacts to the netemul
sandbox, that proxies the custom_artifacts storage capability as a
directory to components in managed realms. This is necessary because
storage is deleted when a component is destroyed, which means any custom
artifacts written by a test component are deleted when the test realm is
torn down. The netemul sandbox creates a new subdirectory for every
component to which it routes custom_artifacts to provide the
per-component isolation that it would otherwise get out of the box from
using a storage capability.

Bug: 378163044
Fixed: 377159657
Test: iperf-test
      netstack_iperf_test
Change-Id: I0db973050737316ec07e64076ada5f2d5c4071d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152492
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>