[flatland] Temporarily disable direct-to-display.
fxb/370839049 aims to relax constraints on image usage by the display
coordinator.  As an intermediate step (fxb/370994314) we want to remove
the abililty of the coordinator to signal release fences directly.

Typically this involves migrating from observing event signals to vsync
stamps, as was done with Carnelian and the Vulkan swapchain.  The same
could have been done here, except that unlike the other cases it would
be throwaway code (since, here, events are used to enforce the
constraints that will soon be relaxed when fxb/370839049 is fixed).

So instead, temporarily disable direct-to-display, and rip out the code
that relies on events.

Bug: 377979329
Fixed: 377967775
Fixed: 42173333
Change-Id: I06a1c9d1ca9ceb03731ecca604bba083e15b496b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153340
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>