[nmfs] Add metric for no events to replay
When the filesystem attempts to reconnect to the socketproxy and there
are no events to replay, such as when the last present network has been
removed, these cases were being incorrectly logged as
"successful_reconnects". We cannot be confident that the reconnect
was successful unless we make a call against the proxy such as
Add/Update/Remove network or SetDefault.

Extract this case into a separate inspect node "no_events_to_replay" to
separate this from being classified as a successful or unsuccessful
reconnect.

Added test to confirm this behavior. This test is separate from the
existing test that validates multi-threaded functionality to assist with
debugging in the case of a while loop hanging.

Fixed: 375233401
Change-Id: I4939b6800eca726a1b7f3ba0458dc1586305503d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144798
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>