[starnix] Carry kmsg log levels to fuchsia syslog severity
Instead of just dumping the message as received, we can carry over the
severity and map to a fuchsia syslog severity. For messages without a
direct mapping we can continue printing the exact level that we received
but map it to a reasonable severity. For example, NOTICE can be an INFO
but we still print <5>.

Bug: 378507649
Change-Id: I1609ae509caf3b5ea9dc8106014d3b1ec0db0d32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153772
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>