[kernel][aarch64] Distinguish captured faults
Not all kinds of data faults on arm can be resolved by the page fault
handler. When not capturing faults this distinction is performed, but
when capturing faults there is not enough information provided to the
caller to know if running the page fault handler can resolve it.

This change exposes additional information to the user_copy routine by
directly setting the return status to indicate whether the fault
information is useful or not.

Change-Id: I9693a201230eef8fd443d781258890b8053844f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1153614
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>