[fuchsia-async] Add close operation to Scope
This is needed in archivist and inspect and probably other places for
handling system shutdown and test scenarios. The basic use case is if
you need to begin graceful shutdown and expect the tasks in your scope
to follow suit because of some coordination happening externally.

Bug: 339724492
Change-Id: I1bce2d106384baa271a43191f5e303c4317ef517
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1154862
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>