[display] Remove //build/config:all_source configs.
//build/config:all_source defines the _ALL_SOURCE macro, which exposes
non-standard features in C headers. The display drivers stack aims to
stick to standard C++ usage, so this config should not be necessary for
our code.

This CL removing the config, reducing the complexity of our build setup,
and removing the possibility that we rely on non-standard features.

This CL introduces a small regression in development ergonomics.
thrd_create_with_name() was a non-standard extension that required the
_ALL_SOURCE macro. This CL replaces thrd_create_with_name() with
thrd_create() in the fake-display and virtio-gpu-display drivers, which
loses the thread names, potentially making debugging a bit more
difficult.

The development regression seems acceptable, because it will be
eliminated when the DFv2 migration is complete. DFv2 drivers use
dispatchers instead of explicitly creating threads.

Test: fx test //src/graphics/display
Change-Id: I5b9e263fa5cd42068b2f2f77c9dae8022c769d57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1152515
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>