[bootstrap][testing] Enable testing support
This change enables running tests through the test framework in a
bringup/bootstrap environment. It does this by including the test
manager and its dependencies, along with the run-test-suite tool which
interacts with it, inside of bootfs in bootstrap builds. This will
allow running bootfs tests as components rather than executables inside
the console.

This change introduces a new testing realm in bootstrap that currently
contains a single entry for driver system tests, which will be used to
run driver power system testing by giving it access to the driver
manager development protocol. We can add additional test collections as
needed in the future.

Change-Id: Ib8ed9d11e4e803b363816992dfd0d671eacbca7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115963
Reviewed-by: Laura Peskin <pesk@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>